+++ mapbox_tilesets/scripts/cli.py
@@ -354,9 +354,43 @@ def job(tileset, job_id, token=None, indent=None):
     is_flag=True,
     help="Will print all tileset information",
 )
+@click.option(
+    "--type",
+    required=False,
+    type=click.Choice(["vector", "raster"]),
+    help="Filter results by tileset type",
+)
+@click.option(
+    "--visibility",
+    required=False,
+    type=click.Choice(["public", "private"]),
+    help="Filter results by visibility",
+)
+@click.option(
+    "--sortby",
+    required=False,
+    type=click.Choice(["created", "modified"]),
+    help="Sort the results by their created or modified timestamps",
+)
+@click.option(
+    "--limit",
+    required=False,
+    type=click.IntRange(1, 500),
+    default=100,
+    help="The maximum number of results to return, from 1 to 500 (default 100)",
+)
 @click.option("--token", "-t", required=False, type=str, help="Mapbox access token")
 @click.option("--indent", type=int, default=None, help="Indent for JSON output")
-def list(username, verbose, token=None, indent=None):
+def list(
+    username,
+    verbose,
+    type=None,
+    visibility=None,
+    sortby=None,
+    limit=None,
+    token=None,
+    indent=None,
+):
     """List all tilesets for an account.
     By default the response is a simple list of tileset IDs.
     If you would like an array of all tileset's information,
@@ -370,6 +404,10 @@ def list(username, verbose, token=None, indent=None):
     url = "{0}/tilesets/v1/{1}?access_token={2}".format(
         mapbox_api, username, mapbox_token
     )
+    url = "{0}&limit={1}".format(url, limit) if limit else url
+    url = "{0}&type={1}".format(url, type) if type else url
+    url = "{0}&visibility={1}".format(url, visibility) if visibility else url
+    url = "{0}&sortby={1}".format(url, sortby) if sortby else url
     r = s.get(url)
     if r.status_code == 200:
         if verbose: