+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 - Pin pytest version
+- Add `--type`, `--visibility`, `--sortby`, and `--limit` options to `list` command
 
 # 1.1.0.dev0 (2020-06-11)
 - `update-recipe` command handles 204 status code in addition to 201 and no longer prints response text